+++ FoodMaintenance/Models/ProductDTO.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string? Name { get; set; }
         public ProductTypeDTO? Type { get; set; }
-        public int MinStockQuantity { get; set; }
+        public float MinStockQuantity { get; set; }
         public UnitOfMeasurementDTO? UnitOfMeasurement { get; set; }
         public bool IsActive { get; set; }
     }
